requirements.txt
fastapi
uvicorn
langchain==0.1.20
langchain-openai
langchain-community
langchain-core
sentence-transformers
chromadb
pypdf
python-dotenv
python-multipart
openai